feat(visualization_simulator):fix#6 Added a pause button
Changed the .js file to add a pause button that transmits to the visualization_simulator file.
Added webserver command for Pause which transmits to visualization_simulator.
Added a bool "paused" to the visualization_simulator that starts as false.
Added a Pause function to visualization_simulator such that if "paused" is false,
and the pause button is pushed, then "paused" becomes true. and vice versa.
Updated Update in visualization_simulator such that if paused is false, it updates, otherwise it does nothing.