feat(bus::Move): feat#12 Skip stops
I figured the only place to go to the next stop was with route.tonextstop.
I found that it was only used in Bus::move function. I made an additional
function that returns true if someone's stop is coming up, or false.
If that returns true, or if the next stop has a passenger at it,
we will unload passengers from the bus and load passengers from the stop.
Otherwise we print that the bus skipped, its id and the stop id and
procede to update the route.