feat(visualization_simulator): fix#15 Added Addlistener, clear ClearListeners
I added AddListener and ClearListeners to the visualization_simulator header file.
I made it doxystyle compliant. Note that the UML we were provided does not align with
my header implementation but I made it that way because that it show it gets called from
my_web_server_command that was provided to us to use.