feat(BusDepot): fix#16 Afternoon is working!
I was getting segfaults and realized that I wasn't ever returning
the bus that the strategy was returning to the depot.
After I fixed that I ran the instance and saw I was only producing small busses.
I realized I was incrementing the state AFTER teh return so it wasn't chaning.
I fixed that and the issue persisted and realized that when I initialize
the state to 0 in the individual strategy it was resetting it evertime so I only
produced small busses. I fixed it and it seems to be working perfectly.