feat(decorator): fix#26 Buses decorated gold on incoming route
I changed the <Bus*> busses_ vectore in vis sim to <IBus*>. I then added
a bool decorated to bus and BusDecorator which represents whether it is
decorated or not. In busses its false, in BusDecorator it returns true.
I also added a functions OutboundComplete which is a bool wheter outroute
IsAtEnd. I used these in Vis Sim's update so that when the outbound is complete
and the bus isn't decorated, we decorate it. I made BusColorDecorator
that inherits from busDecorator and it's GetBusData command gets first
sets the .color aspect to gold before sending it.