fix(observing decorators): fix#28 solved sticky observer problem
For some reason I was only able to access my observers in descending order by id( 1005,1004,...,1000).
After I observed 1000, I couldn't get back to my other busses to observer.
I spoke with Sukhail and he observed I needed to make my register/notify/clear in my BusDecorator.
This was seg faulting and we realized it was because my constructor for ColorDec and IntensityDec were
only writing to their BusToDecorate and not to the one in BusDecorator where the functions were defined.
I changed the construtor and that solved the issue. I also changed it so the Intensity only wraps once at instantiation.