fix(intensity decorator): fix#26 fixed intensity bug
For some reason when I was applying intensity at the factory it wasn't
changing intensity along the route.
I changed it back so it wraps the bus every update with a new intensity and the problem is no longer there.
I'm not sure this is the best solution, but I've had too many issues with this and since it's working, I'm leaving it.